requirements.txt
requests>=2.31.0
Pillow>=10.0.0
qrcode>=7.4.2
pyinstaller>=6.0.0